Update Dart compiler and machine outputs

+++ tests/machine/x/dart/basic_compare.dart
@@ -1,6 +1,6 @@
-var a = 10 - 3;
+var a = (10 - 3 as int);
 
-var b = 2 + 2;
+var b = (2 + 2 as int);
 
 void main() {
   print(a);
