+++ tests/machine/x/dart/binary_precedence.dart
@@ -1,6 +1,6 @@
 void main() {
-  print(1 + 2 * 3);
-  print((1 + 2) * 3);
-  print(2 * 3 + 1);
-  print(2 * (3 + 1));
+  print(((1 + 2 as int) * 3 as int));
+  print((((1 + 2 as int)) * 3 as int));
+  print(((2 * 3 as int) + 1 as int));
+  print((2 * ((3 + 1 as int)) as int));
 }
