+++ tests/machine/x/dart/nested_function.dart
@@ -1,6 +1,6 @@
 int outer(int x) {
   int inner(int y) {
-    return x + y;
+    return (x + y as int);
   }
 
   return inner(5);
