+++ tests/machine/x/dart/partial_application.dart
@@ -1,5 +1,5 @@
 int add(int a, int b) {
-  return a + b;
+  return (a + b as int);
 }
 
 var add5 = (b) => add(5, b);
