+++ tests/machine/x/dart/pure_fold.dart
@@ -1,7 +1,7 @@
 int triple(int x) {
-  return x * 3;
+  return (x * 3 as int);
 }
 
 void main() {
-  print(triple(1 + 2));
+  print(triple((1 + 2 as int)));
 }
