+++ tests/machine/x/dart/tail_recursion.dart
@@ -2,7 +2,7 @@ int sum_rec(int n, int acc) {
   if (n == 0) {
     return acc;
   }
-  return sum_rec(n - 1, acc + n);
+  return sum_rec((n - 1 as int), (acc + n as int));
 }
 
 void main() {
