+++ tests/machine/x/dart/unary_neg.dart
@@ -1,4 +1,4 @@
 void main() {
   print(-3);
-  print(5 + (-2));
+  print((5 + (-2) as int));
 }
